Extract oplog record schema
