+++ Cargo.toml
@@ -19,6 +19,7 @@ members = [
     "crates/repo",
     "crates/review",
     "crates/runtime-bridge",
+    "crates/schema",
     "crates/semantic",
     "crates/state_review",
     "crates/weft-client-shim",
