+++ crates/schema/src/lib.rs
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: Apache-2.0
+//! Pure persisted schema models and codecs shared across Heddle crates.
+
+pub mod op_record;
